requirement.txt
apt-clone==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.2.0
awscli==1.14.44
boto3==1.9.242
botocore==1.12.242
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
Click==7.0
CodeConvert==3.0.2
colorama==0.3.7
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.6
django-admin==2.0.0
django-cors-headers==3.1.0
django-excel-response2==2.0.8
django-extensions==2.2.1
django-models-ext==1.1.9
django-rest-swagger==2.1.2
django-short-url==1.1.4
django-six==1.0.4
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
Flask==1.1.1
furl==2.1.0
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev1
greenlet==0.4.15
httplib2==0.9.2
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
locust==0.0
# Editable install with no version control (locustio==0.11.0)
-e /home/admin1/.local/lib/python3.6/site-packages
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
msgpack-python==0.5.6
mysqlclient==1.4.4
netifaces==0.10.4
oauth==1.0.1
oauth2==1.9.0.post1
oauth2.3==1.5.211
oauthlib==3.1.0
olefile==0.45.1
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==19.2
PAM==0.4.2
pexpect==4.2.1
pika==1.1.0
Pillow==5.1.0
pluggy==0.13.0
protobuf==3.0.0
py==1.8.0
pyasn1==0.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydotplus==2.0.2
pyee==6.0.0
Pygments==2.2.0
pygobject==3.26.1
PyICU==1.9.8
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
pytest==5.2.0
pytest-django==3.5.1
python-apt==1.6.4
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.0
redis==3.3.8
reportlab==3.4.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.2.1
screen==1.0.1
SecretStorage==2.3.1
shortuuid==0.5.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
TimeConvert==1.5.0
tzlocal==2.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.6
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
Werkzeug==0.16.0
xkit==0.0.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.3.2
